Move Deploy from Menu to button as part of header
